結果パターン数の格納はintじゃ足りない

+++ src/main/java/info/shibafu528/shindan4j/Shindan.java
@@ -26,7 +26,7 @@ public interface Shindan extends Serializable{
 
     int getFavoritedCount();
 
-    int getResultPatterns();
+    long getResultPatterns();
 
     boolean isHot();
